Added:Tc-15 1in and 1yard is not equal

+++ src/main/java/com/bridgelabz/Inch.java
@@ -1,11 +1,9 @@
 package com.bridgelabz;
 
-import java.util.Objects;
-
 public class Inch {
     public double value;
 
-    public Inch(int value) {
+    public Inch(double value) {
         this.value=value;
     }
